Added additional cast operation to prevent warning message.
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rulebasedmodels@415 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/C5/src/getnames.c
@@ -630,7 +630,7 @@ void ListAttsUsed()
     {
 	if ( DefOp(D[e]) == OP_ATT )
 	{
-	    Att = (Attribute) DefSVal(D[e]);
+	    Att = (Attribute) (long) DefSVal(D[e]);
 	    if ( ! DefUses[Att] )
 	    {
 		DefUses[Att] = true;
